# use string library to split the string