**Please provide enough information so that others can review your pull request:**

Use string library to split credential to []string. Add checking ensure the result must have only 2 data which is username & password. 

**Commit formatting** 

Use emojis on commit messages so it provides an easy way of identifying the purpose or intention of a commit. Check out the emoji cheatsheet here: https://gitmoji.carloscuesta.me/