@Fenny 2020-10-07
@neil85 thanks for your PR. `strings.Split` will cause an extra allocation which results in slower results.
I added a quick benchmark test https://github.com/gofiber/fiber/pull/895 ( was not included yet ) to make sure this is correct.

```go
// before
Benchmark_Middleware_BasicAuth
Benchmark_Middleware_BasicAuth-16        3809520               311 ns/op              48 B/op          3 allocs/op
Benchmark_Middleware_BasicAuth-16        3870937               311 ns/op              48 B/op          3 allocs/op
Benchmark_Middleware_BasicAuth-16        3870939               316 ns/op              48 B/op          3 allocs/op
Benchmark_Middleware_BasicAuth-16        3857341               310 ns/op              48 B/op          3 allocs/op

// after
Benchmark_Middleware_BasicAuth
Benchmark_Middleware_BasicAuth-16        3100776               385 ns/op              80 B/op          4 allocs/op
Benchmark_Middleware_BasicAuth-16        3108799               385 ns/op              80 B/op          4 allocs/op
Benchmark_Middleware_BasicAuth-16        3116878               386 ns/op              80 B/op          4 allocs/op
Benchmark_Middleware_BasicAuth-16        3116877               384 ns/op              80 B/op          4 allocs/op
```